requirements.txt
appdirs==1.4.3
Django==1.11
django-bootstrap3==8.2.2
django-debug-panel==0.8.3
django-debug-toolbar==1.7
django-debug-toolbar-template-timings==0.7
django-mptt==0.8.7
django-ratelimit==1.0.1
djangorestframework==3.6.2
gunicorn==19.7.1
markdown2==2.3.4
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
sqlparse==0.2.3
